+++ internal/kubernetes/kubernetes.go
@@ -1,9 +1,15 @@
 package kubernetes
 
 import (
+	"context"
+	"slices"
+
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
+	"k8s.io/client-go/util/retry"
 )
 
 func GetClientset(kubeconfigPath string) (*kubernetes.Clientset, error) {
@@ -27,4 +33,16 @@ func GetClientset(kubeconfigPath string) (*kubernetes.Clientset, error) {
 		return nil, err
 	}
 	return clientset, nil
+}
+
+func RemoveNodeTaint(ctx context.Context, cs *kubernetes.Clientset, nodeName, startupTaintKey string) error {
+	return retry.RetryOnConflict(retry.DefaultRetry, func() error {
+		node, err := cs.CoreV1().Nodes().Get(ctx, nodeName, metav1.GetOptions{})
+		if err != nil {
+			return err
+		}
+		node.Spec.Taints = slices.DeleteFunc(node.Spec.Taints, func(t corev1.Taint) bool { return t.Key == startupTaintKey })
+		_, err = cs.CoreV1().Nodes().Update(ctx, node, metav1.UpdateOptions{FieldManager: "spegel"})
+		return err
+	})
 }