+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [#707](https://github.com/spegel-org/spegel/pull/707) Option to remove a taint when spegel is ready.
+
 ### Changed
 
 ### Deprecated